Flytte test-kode fra main til test. Fjerne noen tester som bare tester annen testkode
