+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/innboks/innboksQueries.kt
@@ -8,14 +8,6 @@ import java.sql.PreparedStatement
 import java.sql.ResultSet
 import java.sql.Types
 
-fun Connection.getAllInnboks(): List<Innboks> =
-        prepareStatement("""SELECT * FROM innboks""")
-                .use {
-                    it.executeQuery().list {
-                        toInnboks()
-                    }
-                }
-
 fun Connection.getInnboksById(entityId: Int): Innboks =
         prepareStatement("""SELECT * FROM innboks WHERE id = ?""")
                 .use {
@@ -68,35 +60,7 @@ fun Connection.setInnboksEventerAktivFlag(doneEvents: List<Done>, aktiv: Boolean
     }
 }
 
-fun Connection.getAllInnboksByAktiv(aktiv: Boolean): List<Innboks> =
-        prepareStatement("""SELECT * FROM innboks WHERE aktiv = ?""")
-                .use {
-                    it.setBoolean(1, aktiv)
-                    it.executeQuery().list {
-                        toInnboks()
-                    }
-                }
-
-fun Connection.getInnboksByFodselsnummer(fodselsnummer: String): List<Innboks> =
-        prepareStatement("""SELECT * FROM innboks WHERE fodselsnummer = ?""")
-                .use {
-                    it.setString(1, fodselsnummer)
-                    it.executeQuery().list {
-                        toInnboks()
-                    }
-                }
-
-fun Connection.getInnboksByEventId(eventId: String): Innboks =
-        prepareStatement("""SELECT * FROM innboks WHERE eventId = ?""")
-                .use {
-                    it.setString(1, eventId)
-                    it.executeQuery().singleResult {
-                        toInnboks()
-                    }
-                }
-
-
-private fun ResultSet.toInnboks(): Innboks {
+fun ResultSet.toInnboks(): Innboks {
     return Innboks(
             id = getInt("id"),
             systembruker = getString("systembruker"),
